refactor: isNationalHoliday2, getHolidayName2 から 2 を削除
ロジックベース版がメインになったため、関数名から 2 を削除。
元の Map ベース版（isNationalHoliday, getHolidayName）は削除。
