+++ src/index.ts
@@ -2,8 +2,8 @@
  * デフォルトエントリポイント
  * 祝日判定のみ。祝日名のデータを含まない。
  */
-import { isNationalHoliday2 as isNationalHoliday } from './isNationalHoliday2/index.js';
-import { getHolidayName2 as getHolidayName } from './getHolidayName2/index.js';
+import { isNationalHoliday } from './isNationalHoliday/index.js';
+import { getHolidayName } from './getHolidayName/index.js';
 import { isHoliday } from './isHoliday/index.js';
 import { isBusinessDay } from './isBusinessDay/index.js';
 import { addBusinessDays } from './addBusinessDays/index.js';
